expose the error serialiser for other logger to use
