+++ lib/twiglet/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module Twiglet
-  VERSION = '3.7.2'
+  VERSION = '3.8.0'
 end
